requirements.txt
confluent-kafka
ess-streaming-data-types
